requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI and ML dependencies
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0

# HTTP client for API requests
httpx==0.25.2
requests==2.31.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2

# Logging and monitoring
loguru==0.7.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
aiosqlite==0.19.0

# Development dependencies
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
